+++ ios/ReactNativeSynerise/Modules/RNInjector.m
@@ -3,7 +3,7 @@
 //  react-native-synerise-sdk
 //
 //  Created by Synerise
-//  Copyright © 2018 Synerise. All rights reserved.
+//  Copyright © 2021 Synerise. All rights reserved.
 //
 
 #import "RNInjector.h"