+++ lib/classes/models/Content/Recommendation.d.ts
@@ -1,40 +1,11 @@
 import { BaseModel } from '../BaseModel';
-import { IRecommendationAttribute } from './RecommendationAttribute';
 interface IRecommendation {
-    productRetailerPartNo: string;
-    title: string;
-    brand: string;
-    category: string;
-    description: string;
-    gender: string;
-    color: string;
-    effectivePrice: string;
-    priceCurrency: string;
-    priceValue: string;
-    salePriceValue: string;
-    imageLink: string;
-    link: string;
-    sizes?: Array<string>;
-    additionalImageLinks?: Array<string>;
-    customAttributes?: Array<IRecommendationAttribute>;
+    itemID: string;
+    attributes: Record<string, any>;
 }
 declare class Recommendation extends BaseModel {
-    productRetailerPartNo: string;
-    title: string;
-    brand: string;
-    category: string;
-    description: string;
-    gender: string;
-    color: string;
-    effectivePrice: string;
-    priceCurrency: string;
-    priceValue: string;
-    salePriceValue: string;
-    imageLink: string;
-    link: string;
-    sizes?: Array<string>;
-    additionalImageLinks?: Array<string>;
-    customAttributes?: Array<IRecommendationAttribute>;
+    itemID: string;
+    attributes: Record<string, any>;
     constructor(modelObject: IRecommendation);
 }
 export { IRecommendation, Recommendation };