+++ lib/classes/models/Content/ScreenViewAudience.d.ts
@@ -1,0 +1,11 @@
+import { BaseModel } from '../BaseModel';
+interface IScreenViewAudience {
+    IDs?: Array<String>;
+    query?: string;
+}
+declare class ScreenViewAudience extends BaseModel {
+    IDs?: Array<String>;
+    query?: string;
+    constructor(modelObject: IScreenViewAudience);
+}
+export { IScreenViewAudience, ScreenViewAudience };